Remove old unused files and move some other
