+++ frontend/task/task.scss
@@ -38,7 +38,7 @@
 
 @import "./layout/layout.scss";
 @import "./controls.scss";
-@import "./documentation.scss";
+@import "./documentation/documentation.scss";
 @import "./task_tests.scss";
 @import "./levels.scss";
 @import "./blocks/blocks.scss";
